OSSM-6774 Use the existing kube clients to create metrics resources (#1827)
Previously, during startup, the operator instantiated a separate kube client and a separate discovery client to create the Service and ServiceMonitor for metrics. Now, the existing clients provided by the controller manager are used to create these two resources.

However, since these clients are only initialized when the manager is started, the code that creates the two metrics resources is now invoked by a runnable registered with the manager via `mgr.Add()`.